+++ NotificarePushUIKit/Sources/NotificarePushUI+NotificareNotification.swift
@@ -13,6 +13,17 @@ public extension NotificareNotification {
         switch type {
         case .none, .passbook, .alert, .rate, .store, .urlScheme, .inAppBrowser:
             return false
+
+        case .urlResolver:
+            let result = NotificationUrlResolver.resolve(self)
+
+            switch result {
+            case .none, .urlScheme, .inAppBrowser:
+                return false
+            case .webView:
+                return true
+            }
+
         default:
             return true
         }