+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 3.10.0
+
+- Add support for the URLResolver notification type
+- Fix anonymous device registration
+
 ## 3.9.1
 
 - Preload images before showing in-app messages